Refactor mappers
