+++ src/main/java/com/msn/msncars/car/make/MakeDTO.java
@@ -1,4 +1,6 @@
-package com.msn.msncars.car;
+package com.msn.msncars.car.make;
+
+import com.msn.msncars.car.model.ModelDTO;
 
 import java.util.List;
 public record MakeDTO(
