+++ src/main/java/com/msn/msncars/car/make/MakeRepository.java
@@ -1,4 +1,4 @@
-package com.msn.msncars.car;
+package com.msn.msncars.car.make;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
