+++ src/main/java/com/msn/msncars/car/model/ModelDTO.java
@@ -1,5 +1,6 @@
-package com.msn.msncars.car;
+package com.msn.msncars.car.model;
 
+import com.msn.msncars.car.make.MakeSlimDTO;
 public record ModelDTO(
         Long id,
         String name,
