+++ src/main/java/com/msn/msncars/car/model/ModelMapper.java
@@ -1,0 +1,10 @@
+package com.msn.msncars.car.model;
+
+
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface ModelMapper {
+    Model fromDTO(ModelDTO modelDTO);
+    ModelDTO toDTO(Model model);
+}
