+++ src/main/java/com/msn/msncars/car/model/ModelRepository.java
@@ -1,4 +1,4 @@
-package com.msn.msncars.car;
+package com.msn.msncars.car.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
